pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["nikolay1-97 <nikolay111_connect@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
alembic = "^1.13.1"
sqlalchemy = "2.0.25"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.3"
asyncpg = "0.29.0"
asyncio = "^3.4.3"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"
httpx = "^0.27.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
pytest-cov = "^5.0.0"
fastapi-redis-cache = "^0.2.5"
aioredis = "^2.0.1"
openpyxl = "^3.1.2"
celery = "^5.3.6"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
gunicorn = "^22.0.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode="auto"
pythonpath = [
  ".", "app",
]
